+++ changelog.md
@@ -5,9 +5,11 @@
 - 增强移动端添加设置按钮的适应性
 - 适配思源笔记3.7.0版本新语言代码
 - 优化打字机模式
+- 在主题设置中关闭打字机模式开关后正确禁用打字机模式
 - Enhance the adaptability of the add settings button on mobile
 - Adapt to the new language code of SiYuan Note 3.7.0 version
 - Optimize typewriter mode
+- Disable typewriter mode properly when the typewriter mode switch is turned off in the theme settings
 
 ## v3.0.0